WIP: Use package-specific archive and subdir
When a repo contains multiple released packages, each individual package
can be downloaded as a standalone repo from the release repository.

This commit prefers downloading from the release repository and downloads
a package-only archive as source archive.

This source archive contains (from my understanding) only one directory
containing the corresponding package, so we can use this as the _dir directly.

This is implemented quite hacky ATM and should serve as a PoC only.